--- local/yunita/db/access.php
@@ -1,14 +1,0 @@
-<?php
-    $capabilities = array(
-        
-        'local/yunita:add' => array(
-            
-            'captype' => 'write',
-            'contextlevel' => CONTEXT_SYSTEM,
-            'archetypes' => array( // The roles that you want to allow
-                'teacher' => CAP_ALLOW,
-                'manager' => CAP_ALLOW
-            ),
-        ),
-    );
-?>
