--- local/yunita/lang/en/local_yunita.php
@@ -1,3 +1,0 @@
-<?php
-    $string['pluginname'] = 'YUNITA Plug-in';
-?>
