--- local/yunita/version.php
@@ -1,6 +1,0 @@
-<?php
-    $plugin->version = 2016012300;
-    $plugin->requires = 2011033006;
-    $plugin->component = 'local_yunita';
-    $plugin->maturity = MATURITY_ALPHA;
-?>
